Changes for Mary 21-03-2016
1 On the home page Susan McPhee’s (Treasurer) email address is correct
in the left hand margin, however down the bottom of that page the email
address is not correct.
2 The Vice President has been changed therefore Denise Maidment needs to
be omitted and the name Robin McPhee inserted instead. His email address
is turbot@ihug.co.nz.
3 Can you delete the membership form and insert the new one (attached)
as Susan McPhee’s email address had to be changed on that form too, so I
have attached a new form.